[feat] Item/Resource Object Pool 삭제
