--- Assets/DEV/YJE/Item.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c3e61a565ac91c94e8587f46f83f1002
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
