--- Assets/DEV/YJE/ItemObj.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9342fe136426c2e4084739fcd694da34
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
