--- Assets/DEV/YJE/ItemPool.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: adaef141468e4f14ab044f4a44c886b1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
