--- Assets/DEV/YJE/OreSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bab8c5f00b664fb44b24026e5d07fbd5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
