+++ Assets/DEV/YJE/ResourceController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 33409cb3f00f58644b1c58fee57b0ba9
+guid: 0ab187da7be87ab49987baea8943a776
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
