+++ Assets/DEV/YJE/Scripts/Delete.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: a3563da7bdf73a2449a1574de7b5b585
+guid: 14ed581423a067a49b9b6dd9974ef735
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
