--- Assets/DEV/YJE/Scripts/ResourceController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c545a781b890964b90d0fec1b73f77b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
