--- Assets/DEV/YJE/WoodSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cacaac6353b5f8f41bea3dea8d7dcc26
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
